feat: renamed configure command to config for consistency
